refactor(pagination): calculate first-item and last-item in parent
